test вакансия /backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
graphene==3.3
graphene-sqlalchemy==2.3.0
flask-graphql==2.0.1
python-dotenv==1.0.0
alembic==1.12.1
pytest==7.4.3
